requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.4.0
billiard==3.6.4.0
boto3==1.17.52
botocore==1.20.52
celery==5.1.2
certifi==2019.6.16
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.4.2
Django==3.2.6
django-allauth==0.42.0
django-autoslug==1.9.3
django-cors-headers==3.7.0
django-crispy-forms==1.14.0
django-djconfig==0.5.3
django-filter==2.4.0
django-formtools==2.3
django-import-export==2.5.0
django-mathfilters==0.4.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.2.0
django-rest-swagger==2.2.0
django-storages==1.11.1
django-templated-mail==1.1.1
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
gunicorn==20.1.0
h11==0.8.1
h2==3.2.0
hpack==3.0.0
http3==0.6.7
hyperframe==5.2.0
idna==2.8
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
jsonfield==3.1.0
jupyter-core==4.10.0
kombu==5.1.0
MarkupPy==1.14
MarkupSafe==2.0.1
nest-asyncio==1.5.5
oauthlib==3.2.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.9
packaging==21.0
parso==0.8.3
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
pure-eval==0.2.2
pycparser==2.21
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.1
python3-openid==3.2.0
pytz==2019.3
pywin32==304
PyYAML==6.0
pyzmq==23.0.0
redis==3.5.3
requests==2.22.0
requests-async==0.6.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.3.7
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.3.0
tablib==0.14.0
tornado==6.1
traitlets==5.2.1.post0
tzdata==2022.1
uritemplate==3.0.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
whitenoise==3.3.1
xlrd==2.0.1
xlwt==1.3.0
pytest-django==4.5.2
